+++ frameioclient/lib/version.py
@@ -1,0 +1,9 @@
+try:
+    from importlib import metadata
+except ImportError:
+    # Running on pre-3.8 Python; use importlib-metadata package
+    import importlib_metadata as metadata
+
+class ClientVersion:
+    def version():
+        return metadata.version('frameioclient')
